requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
plotly>=5.0.0
duckduckgo-search>=6.0.0
python-dotenv>=1.0.0
